add frames_utils.py add function for combine frames
