+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,6 +1,49 @@
 <component name="InspectionProjectProfileManager">
   <profile version="1.0">
     <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="35">
+            <item index="0" class="java.lang.String" itemvalue="graphql-relay" />
+            <item index="1" class="java.lang.String" itemvalue="PyYAML" />
+            <item index="2" class="java.lang.String" itemvalue="python-dotenv" />
+            <item index="3" class="java.lang.String" itemvalue="h11" />
+            <item index="4" class="java.lang.String" itemvalue="promise" />
+            <item index="5" class="java.lang.String" itemvalue="MarkupSafe" />
+            <item index="6" class="java.lang.String" itemvalue="requests" />
+            <item index="7" class="java.lang.String" itemvalue="Jinja2" />
+            <item index="8" class="java.lang.String" itemvalue="graphene" />
+            <item index="9" class="java.lang.String" itemvalue="httptools" />
+            <item index="10" class="java.lang.String" itemvalue="starlette" />
+            <item index="11" class="java.lang.String" itemvalue="certifi" />
+            <item index="12" class="java.lang.String" itemvalue="urllib3" />
+            <item index="13" class="java.lang.String" itemvalue="uvicorn" />
+            <item index="14" class="java.lang.String" itemvalue="itsdangerous" />
+            <item index="15" class="java.lang.String" itemvalue="uvloop" />
+            <item index="16" class="java.lang.String" itemvalue="websockets" />
+            <item index="17" class="java.lang.String" itemvalue="email-validator" />
+            <item index="18" class="java.lang.String" itemvalue="dnspython" />
+            <item index="19" class="java.lang.String" itemvalue="pydantic" />
+            <item index="20" class="java.lang.String" itemvalue="async-exit-stack" />
+            <item index="21" class="java.lang.String" itemvalue="aniso8601" />
+            <item index="22" class="java.lang.String" itemvalue="Rx" />
+            <item index="23" class="java.lang.String" itemvalue="graphql-core" />
+            <item index="24" class="java.lang.String" itemvalue="aiofiles" />
+            <item index="25" class="java.lang.String" itemvalue="orjson" />
+            <item index="26" class="java.lang.String" itemvalue="typing-extensions" />
+            <item index="27" class="java.lang.String" itemvalue="python-multipart" />
+            <item index="28" class="java.lang.String" itemvalue="click" />
+            <item index="29" class="java.lang.String" itemvalue="chardet" />
+            <item index="30" class="java.lang.String" itemvalue="fastapi" />
+            <item index="31" class="java.lang.String" itemvalue="watchgod" />
+            <item index="32" class="java.lang.String" itemvalue="async-generator" />
+            <item index="33" class="java.lang.String" itemvalue="idna" />
+            <item index="34" class="java.lang.String" itemvalue="ujson" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
     <inspection_tool class="PyPep8NamingInspection" enabled="true" level="WEAK WARNING" enabled_by_default="true">
       <option name="ignoredErrors">
         <list>
